# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Tappix Loyalty Platform API
  version: 1.0.0
  description: Multi-tenant REST API for Tappix business, customer and super-admin applications.
servers:
  - url: http://localhost:8080/api/v1
security: [{bearerAuth: []}]
paths:
  /auth/login:
    post:
      security: []
      summary: Business or super-admin login
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/LoginInput'}}}}
      responses: {'200': {$ref: '#/components/responses/Success'}, '401': {$ref: '#/components/responses/Error'}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
...
  schemas:
    LoginInput:
      type: object
      required: [email, password]
      properties: {email: {type: string, format: email}, password: {type: string, format: password}}
...
    Envelope:
      type: object
      required: [success]
      properties: {success: {type: boolean}, data: {}, error: {$ref: '#/components/schemas/APIError'}}
    APIError:
      type: object
      properties: {code: {type: string}, message: {type: string}}
  responses:
    Success: {description: Successful envelope, content: {application/json: {schema: {$ref: '#/components/schemas/Envelope'}}}}
    Error: {description: Error envelope, content: {application/json: {schema: {$ref: '#/components/schemas/Envelope'}}}}